win64 fixes (partial)
cherry-picked from 2f734ae2317060edb83021f17cffc966c435ad7b

+++ library/include/VersionInfo.h
@@ -53,7 +53,7 @@ namespace DFHack
         std::map <std::string, uintptr_t> Addresses;
         std::map <std::string, uintptr_t> VTables;
         uintptr_t base;
-        int rebase_delta;
+        uintptr_t rebase_delta;
         std::string version;
         OSType OS;
     public:
